Partial insert should not skip "not null" columns that have null values
When "not null" column was omitted from insert statement in MySQL non-strict mode, MySQL silently puts something instead of raising an error,
for example, 0 for an integer column, '' for a varchar column, '0000-00-00' for a date column.
This fix brings AR 3.2 behaviour back, so as to raise null constraint error when nil was given to a "not null" column.